+++ src/components/StockSearchInput/StockSearchInput.tsx
@@ -102,14 +102,13 @@ export default function StockSearchInput() {
           currency: selectedStockDetails.currency,
         };
         const timeStamp = Date.now();
-        const currentTime = new Date();
-        const hours = currentTime.getHours().toString().padStart(2, "0");
-        const minutes = currentTime.getMinutes().toString().padStart(2, "0");
-        console.log(Object.keys(stockDailyData)[0]);
 
         addToRecentlySeenStocks({
           lastUpdated: timeStamp,
-          quoteLastUpdate: `${hours} : ${minutes}`,
+          quoteLastUpdate: new Date().toLocaleTimeString([], {
+            hour: "2-digit",
+            minute: "2-digit",
+          }),
           historicalDataLastUpdate: Object.keys(stockDailyData)[0],
           stockOverview: stockOverview,
           stockDetails: selectedStockDetails,
