+++ src/helpers/fetchHistoricalData.ts
@@ -1,3 +1,4 @@
+//delete?
 import axios from "axios";
 const ALPHA_VANTAGE_API_KEY = import.meta.env.VITE_ALPHA_VANTAGE_API_KEY;
