+++ src/helpers/fetchQuote.ts
@@ -1,3 +1,4 @@
+//maybe delete
 import axios from "axios";
 const FINNHUB_API_KEY = import.meta.env.VITE_FINNHUB_API_KEY;
 const FINNHUB_BASE_URL = "https://finnhub.io/api/v1";
