+++ src/pages/NoData.tsx
@@ -5,7 +5,7 @@ export default function NoData() {
   const { containersColors } = useStocksContext();
   return (
     <section className={`noData_container ${containersColors}`}>
-      <h1>No Data Avalible </h1>
+      <h1>No Data Available </h1>
     </section>
   );
 }
